sdk: split query_verbs into module directory (6/9: extract flow)
Move turn_span_tree, session_span_trees, flow_graph, context_delta, and
helpers to query_verbs/flow.rs.
Visibility: bucket_subagents_per_turn, duration_to_since_iso -> pub(crate)
(used directly from tests module); add explicit imports in tests.rs.
